Fix lasers not rendering

+++ src/main/java/mekanism/common/particle/LaserParticleData.java
@@ -12,16 +12,16 @@ import net.minecraft.core.particles.ParticleType;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 
-public record LaserParticleData(Direction direction, double distance, float energyScale) implements ParticleOptions {
+public record LaserParticleData(Direction direction, float distance, float energyScale) implements ParticleOptions {
 
     public static final MapCodec<LaserParticleData> CODEC = RecordCodecBuilder.mapCodec(val -> val.group(
           Direction.CODEC.fieldOf(SerializationConstants.DIRECTION).forGetter(data -> data.direction),
-          Codec.DOUBLE.fieldOf(SerializationConstants.DISTANCE).forGetter(data -> data.distance),
+          Codec.FLOAT.fieldOf(SerializationConstants.DISTANCE).forGetter(data -> data.distance),
           Codec.FLOAT.fieldOf(SerializationConstants.ENERGY).forGetter(data -> data.energyScale)
     ).apply(val, LaserParticleData::new));
     public static final StreamCodec<ByteBuf, LaserParticleData> STREAM_CODEC = StreamCodec.composite(
           Direction.STREAM_CODEC, LaserParticleData::direction,
-          ByteBufCodecs.DOUBLE, LaserParticleData::distance,
+          ByteBufCodecs.FLOAT, LaserParticleData::distance,
           ByteBufCodecs.FLOAT, LaserParticleData::energyScale,
           LaserParticleData::new
     );
